Update connect-wallet.html
Added in two different scripts to initialize Moralis, Metamask & WalletConnect in the head of the HTML file. 

Added individual IDs to each one of the buttons (Metamask & WalletConnect). Did not add IDs to Formatic and Coinbase wallets as they are implemented into WalletConnect. Each ID calls the login functions associated with the script.js file.